+++ src/lib.rs
@@ -67,16 +67,20 @@
 
 mod bounds;
 mod intervals;
+mod iterator;
 mod multi_intervals;
 mod nothing_between;
+mod step;
 mod tests;
 
 #[cfg(feature = "serde")]
 mod serde;
 
 pub use crate::intervals::{Interval, ParseError};
+pub use crate::iterator::IntervalIterator;
 pub use crate::multi_intervals::MultiInterval;
 pub use crate::nothing_between::NothingBetween;
+pub use crate::step::Step;
 
 /// This macro lets you create intervals with a syntax closer to what Postgresql
 /// provides.